Finish form style and client-side validation through JavaScript.
Created function passwordMatch to compare both password input values
and put a text cue to an empty <p>  depending on if the passwords match
or not. They will be colored as red or white accordingly. The function
will trigger whenever there is a change in the input field.

Created function styleInvalidLabels to listen to the "focusout" event
on all the inputs, and color its associated labels red whenever the
input field text is also red, or grey otherwise.
 --> I selected this approach to color the labels because they are the
 input's previous sibling, and they cannot be easily targeted with css.
 I thought of using the "fieldset:invalid" combination to select them
 but this method did not work for me because all the camps are required,
 so the labels would appear red from the start. This combination does
 not seem to work with the ":user-invalid" pseudo-class, which would
 have solved the issue.

Created function noSubmit to prevent the default behaviour of the
submit button whenever it is clicked but the passwords are not matching
(even if all the input camps are correct according to the pattern
validation tests).

Used CSS pseudo-classes and "generate content" to place textual cues
about how the inputs are meant to be answered, whenever the user focuses
them.

Used the title attribute within inputs to modify the base tips from the
browser validation.

Used the pattern attribute with according RegEx to create the client-side
validation for the different input fields.