Rename and Update function styleInvalidLabels (now styleInputElements)
Renamed and updated the function:

Now when an input is focused-out it will update both labels and the cue
paragraphs.
--> Labels will be colored red when the input content is invalid and
white when it is valid.
--> Cue paragraphs will be colored red when the input is invalid and
transparent when it is valid.

The function will also trigger when an input is focused-in updating
cue paragraphs.
--> Cue paragraphs will be colored white whenever they are focused-in
and remain that way but only if they have a value, remaining transparent
if not.

The validation tests for the function are based on the input element's
text-color, which in turn is validated by the RegEx used in HTML pattern
attributes via CSS.