Finish function submit (previously named noSubmit)
Function noSubmit was unfinished. Now it contemplates two different
scenarios when the user uses the submit button:

--> If all the camps have been filled and are valid, the page displays
an alert congratulating the user for creating the account and submits
the form (even though the data goes anywhere).
This feature is accomplished by testing all the inputs color is white
(which means the input is valid) by using the computedStyle method.

--> If all the camps have been filled and are valid, but the passwords
do not match, it cancels the buttons submit behaviour and displays an
alert explaining the user needs to fix the passwords.

In case any of the input fields has not been filled or its content is
not valid, the browser's built-in validation will still act and prevent
the button's behaviour, displaying also an alert on the invalid fields.