Improve look in small desktop sizes.
Form was overflowing from its box within small desktop sizes.
html height was deleted to allow all content to overflow the desktop
view if it doesn't fit. Now users can scroll down an the content keeps
within the box.

Fixed also the brand picture positioning, so it is better aligned within
its div in small desktops.

Gave the main background image's cite a position relative to put it
within the screen, as with the new changes it was cut from the screen
even in big desktop screens.